pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bamboo-mcp-services"
version = "1.0.0"
description = "Bamboo MCP Services"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=0.10",
    "requests>=2.31",
    "pyyaml>=6.0",
    "pandas>=2.0",
]

[project.optional-dependencies]
dev = [
    "flake8",
    "pytest",
    "pytest-cov",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
bamboo-document-monitor = "bamboo_mcp_services.agents.document_monitor_agent.cli:main"
bamboo-github-sync = "bamboo_mcp_services.agents.github_doc_sync_agent.cli:main"
bamboo-ingestion = "bamboo_mcp_services.agents.ingestion_agent.cli:main"
bamboo-cric = "bamboo_mcp_services.agents.cric_agent.cli:main"
bamboo-dummy = "bamboo_mcp_services.agents.dummy_agent.cli:main"
